// ButorRevolutionWebAPI/src/frontend/web/projects/butor-shared-lib/src/lib/models/user.basic.model.ts
export interface IUserBasicInfo {
    name: string;
    phone?: string;
    email: string;
    countryId?: number;
    postCode?: number;
    city?: string;
    address?: string;
    languageId?: number;
    fileName?: string;
    containerName?: string;
    currentPassword?: string;
    newPassword?: string;
}